Reworked Launch part
