+++ fe1-web/parts/Launch.tsx
@@ -1,0 +1,109 @@
+import React, { useState } from 'react';
+import {
+  StyleSheet, View, TextInput, TextStyle, ViewStyle,
+} from 'react-native';
+import { Spacing, Typography } from 'styles';
+
+import STRINGS from 'res/strings';
+import PROPS_TYPE from 'res/Props';
+import { requestCreateLao } from 'network/MessageApi';
+import { dispatch } from 'store';
+import { getNetworkManager } from 'network';
+import { JsonRpcResponse } from 'model/network';
+import WideButtonView from 'components/WideButtonView';
+import TextBlock from 'components/TextBlock';
+import PropTypes from 'prop-types';
+import styleContainer from 'styles/stylesheets/container';
+
+/**
+ * Manage the Launch screen: a description string, a LAO name text input, a launch LAO button,
+ * and cancel button
+ *
+ * The Launch button does nothing
+ * The cancel button clear the LAO name field and redirect to the Home screen
+ *
+ * TODO implement the launch button action
+*/
+const styles = StyleSheet.create({
+  textInput: {
+    ...Typography.base,
+    borderBottomWidth: 2,
+    marginVertical: Spacing.s,
+    marginHorizontal: Spacing.xl,
+  } as TextStyle,
+  viewTop: {
+    justifyContent: 'flex-start',
+  } as ViewStyle,
+  viewBottom: {
+    justifyContent: 'flex-end',
+  } as ViewStyle,
+});
+
+const Launch = ({ navigation }: IPropTypes) => {
+  const [inputLaoName, setInputLaoName] = useState('');
+
+  const onButtonLaunchPress = (laoName: string) => {
+    if (laoName) {
+      requestCreateLao(laoName);
+    } else {
+      console.error('empty lao name');
+    }
+  };
+
+  const onTestOpenConnection = () => {
+    const nc = getNetworkManager().connect('127.0.0.1');
+    nc.setResponseHandler((m: JsonRpcResponse) => {
+      console.info('Handling the json-rpc response : ', m);
+    });
+  };
+
+  const onTestClearStorage = () => {
+    dispatch({ type: 'CLEAR_STORAGE', value: {} });
+  };
+
+  const cancelAction = () => {
+    setInputLaoName('');
+    navigation.navigate('Home');
+  };
+
+  return (
+    <View style={styleContainer.flex}>
+      <View style={styles.viewTop}>
+        <TextBlock text={STRINGS.launch_description} />
+        <TextInput
+          style={styles.textInput}
+          placeholder={STRINGS.launch_organization_name}
+          onChangeText={(input: string) => setInputLaoName(input)}
+          defaultValue={inputLaoName}
+        />
+      </View>
+      <View style={styles.viewBottom}>
+        <WideButtonView
+          title={STRINGS.launch_button_launch}
+          onPress={() => onButtonLaunchPress(inputLaoName)}
+        />
+        <WideButtonView
+          title="[TEST] Connect to LocalMockServer.ts (use 'npm run startServer')"
+          onPress={onTestOpenConnection}
+        />
+        <WideButtonView
+          title="[TEST] Clear (persistent) storage"
+          onPress={onTestClearStorage}
+        />
+        <WideButtonView
+          title={STRINGS.general_button_cancel}
+          onPress={cancelAction}
+        />
+      </View>
+    </View>
+  );
+};
+
+const propTypes = {
+  navigation: PROPS_TYPE.navigation.isRequired,
+};
+Launch.propTypes = propTypes;
+
+type IPropTypes = PropTypes.InferProps<typeof propTypes>;
+
+export default Launch;
